added shadowbox result, output hdr as float32, robertson output log curve

+++ imageIO.py
@@ -3,6 +3,7 @@ import cv2, exifread
 import os
 import math
 
+
 def read_images(image_dir):
     paths = [os.path.join(image_dir, file) for file in os.listdir(image_dir) if os.path.isfile(os.path.join(image_dir, file))]
     LDR_images = []
@@ -41,4 +42,4 @@ def transform_exif_fraction_to_float(fraction):
     return numbers[0]/numbers[1]
 
 def save_HDR_images(image, output):
-    cv2.imwrite(output, image)
+    cv2.imwrite(output, image.astype(np.float32))
